solve dependency issues, change total return type to int and refactor test, unfortunately I have to remove Spring Fox Dependency due to the issues with other dependencies

+++ server/src/main/java/com/sparta/domain/repository/DataRepository.java
@@ -4,7 +4,7 @@ public interface DataRepository<T> {
 
     public void save(String key, T value);
 
-    public long countByProvider(String provider);
+    public int countByProvider(String provider);
 
 	public void deleteAll();
 }
